feat: added language translations to helpcenter topics
