+++ src/modules/help-center/help-center.service.ts
@@ -8,6 +8,7 @@ import { SearchHelpCenterDto } from './dto/search-help-center.dto';
 import { REQUEST_SUCCESSFUL, QUESTION_ALREADY_EXISTS, USER_NOT_FOUND } from '../../helpers/SystemMessages';
 import { CustomHttpException } from '../../helpers/custom-http-filter';
 import { User } from '../user/entities/user.entity';
+import { TextService } from '../translation/translation.service';
 
 @Injectable()
 export class HelpCenterService {
@@ -15,10 +16,15 @@ export class HelpCenterService {
     @InjectRepository(HelpCenterEntity)
     private readonly helpCenterRepository: Repository<HelpCenterEntity>,
     @InjectRepository(User)
-    private userRepository: Repository<User>
+    private userRepository: Repository<User>,
+    private readonly textService: TextService
   ) {}
 
-  async create(createHelpCenterDto: CreateHelpCenterDto, user: User) {
+  private async translateContent(content: string, lang: string) {
+    return this.textService.translateText(content, lang);
+  }
+
+  async create(createHelpCenterDto: CreateHelpCenterDto, user: User, language: string = 'en') {
     const existingTopic = await this.helpCenterRepository.findOne({
       where: { title: createHelpCenterDto.title },
     });
@@ -36,8 +42,14 @@ export class HelpCenterService {
       throw new CustomHttpException(USER_NOT_FOUND, HttpStatus.NOT_FOUND);
     }
 
+    let translatedContent = createHelpCenterDto.content;
+    if (language && language !== 'en') {
+      translatedContent = await this.translateContent(createHelpCenterDto.content, language);
+    }
+
     const helpCenter = this.helpCenterRepository.create({
       ...createHelpCenterDto,
+      content: translatedContent,
       author: `${fullUser.first_name} ${fullUser.last_name}`,
     });
     const newEntity = await this.helpCenterRepository.save(helpCenter);
@@ -49,11 +61,18 @@ export class HelpCenterService {
     };
   }
 
-  async findAll(): Promise<any> {
+  async findAll(language?: string): Promise<any> {
     const centres = await this.helpCenterRepository.find();
+    const translatedhCTopics = await Promise.all(
+      centres.map(async topic => {
+        topic.title = await this.translateContent(topic.title, language);
+        topic.content = await this.translateContent(topic.content, language);
+        return topic;
+      })
+    );
 
     return {
-      data: centres,
+      data: translatedhCTopics,
       status_code: HttpStatus.OK,
       message: REQUEST_SUCCESSFUL,
     };
